requirements.txt
pydub==0.24.1
SpeechRecognition==3.8.1
streamlit==0.61.0
Pillow==8.0.0
